Fix SSLInterceptor NPE
The spec states returning a non-null (possibly empty) array of issuers
and whilst a null works on the default it causes a NPE in some
implementations.
Whilst I can’t find doc that clearly states it behaviour suggests that
the empty array is a special case that doesn’t mean trust none, but
trust all.